+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # TelegramBotAPI changelog
 
+## 0.38.15
+
+* `Common`:
+    * `Version`:
+        * `MicroUtils`: `0.9.20` -> `0.9.24`
+* `Core`:
+    * Fixes in `MessageContent#serializationModule`
+* `BehaviourBuilder`:
+    * Add triggers for `DataCallbackQuery` and subtypes with regex checking of data
+
 ## 0.38.14
 
 __This update contains including of [Telegram Bot API 6.0](https://core.telegram.org/bots/api-changelog#april-16-2022)__